fix: Refactor: Convert repository structure from technical layers to Domain-Driven Design (DDD) (#9108)
