+++ src/components/common/WorkflowActionsDropdown.vue
@@ -9,7 +9,7 @@ import { useI18n } from 'vue-i18n'
 
 import WorkflowActionsList from '@/components/common/WorkflowActionsList.vue'
 import Button from '@/components/ui/button/Button.vue'
-import { useNewMenuItemIndicator } from '@/composables/useNewMenuItemIndicator'
+import { useNewMenuItemIndicator } from '@/components/common/useNewMenuItemIndicator'
 import { useWorkflowActionsMenu } from '@/composables/useWorkflowActionsMenu'
 import { useTelemetry } from '@/platform/telemetry'
 import { useCanvasStore } from '@/renderer/core/canvas/canvasStore'
