+++ src/components/curve/useCurveEditor.ts
