+++ src/components/graph/GraphCanvas.vue
@@ -148,9 +148,9 @@ import type { VueNodeData } from '@/composables/graph/useGraphNodeManager'
 import { useVueNodeLifecycle } from '@/composables/graph/useVueNodeLifecycle'
 import { useNodeBadge } from '@/composables/node/useNodeBadge'
 import { useCanvasDrop } from '@/composables/useCanvasDrop'
-import { useContextMenuTranslation } from '@/composables/useContextMenuTranslation'
+import { useContextMenuTranslation } from '@/components/graph/useContextMenuTranslation'
 import { useCopy } from '@/composables/useCopy'
-import { useGlobalLitegraph } from '@/composables/useGlobalLitegraph'
+import { useGlobalLitegraph } from '@/components/graph/useGlobalLitegraph'
 import { usePaste } from '@/composables/usePaste'
 import { useVueFeatureFlags } from '@/composables/useVueFeatureFlags'
 import { LiteGraph } from '@/lib/litegraph/src/litegraph'
