+++ src/components/graph/GraphCanvasMenu.vue
@@ -90,7 +90,7 @@ import { computed, onBeforeUnmount, onMounted } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 import Button from '@/components/ui/button/Button.vue'
-import { useZoomControls } from '@/composables/useZoomControls'
+import { useZoomControls } from '@/components/graph/useZoomControls'
 import { LiteGraph } from '@/lib/litegraph/src/litegraph'
 import { useSettingStore } from '@/platform/settings/settingStore'
 import { useTelemetry } from '@/platform/telemetry'
