+++ src/renderer/extensions/vueNodes/widgets/composables/useTransformCompatOverlayProps.ts
