+++ src/workbench/extensions/manager/composables/useServerLogs.ts
